fix: re-formatted code for better readability
