+++ .env.example
@@ -1,6 +1,8 @@
 HOST=
 PORT=
 NODE_ENV=
+DATABASE_URL_TEST=
+DATABASE_URL_DEV=
 DATABASE_URL=
 JWT_KEY=
 EMAIL_USER=
