--- src/utils/responseHelpers.js
@@ -1,20 +1,0 @@
-function sendErrorResponse(response, status, message, errors) {
-    response.status(status).json({
-        status: status,
-        message: message,
-        errors: errors
-    });
-}
-
-function sendSuccessResponse(response, status, message, data) {
-    response.status(status).json({
-        status: status,
-        message: message,
-        ...data
-    });
-}
-
-module.exports = {
-    sendErrorResponse,
-    sendSuccessResponse
-};
